front: editor: remove dependencies cycles
