+++ front/src/applications/editor/context.ts
@@ -1,6 +1,5 @@
 import { Context, createContext } from 'react';
-
-import { EditorContextType } from './tools/types';
+import { EditorContextType } from './tools/editorContextTypes';
 
 const EditorContext = createContext<EditorContextType | null>(null) as Context<EditorContextType>;
