+++ front/src/applications/editor/tools/toolTypes.ts
@@ -1,0 +1,12 @@
+enum TOOL_TYPES {
+  SELECTION = 'select-items',
+  TRACK_EDITION = 'track-edition',
+  SPEED_SECTION_EDITION = 'speed-edition',
+  SWITCH_EDITION = 'switch-edition',
+  SIGNAL_EDITION = 'signal-edition',
+  DETECTOR_EDITION = 'detector-edition',
+  BUFFER_STOP_EDITION = 'buffer-stop-edition',
+  ROUTE_EDITION = 'route-edition',
+}
+
+export default TOOL_TYPES;
