+++ front/src/applications/editor/tools/trackEdition/tool.ts
@@ -14,7 +14,6 @@ import nearestPointOnLine, { NearestPointOnLine } from '@turf/nearest-point-on-l
 import getNearestPoint from '@turf/nearest-point';
 import { featureCollection } from '@turf/helpers';
 
-import { DEFAULT_COMMON_TOOL_STATE, Tool } from '../types';
 import { getMapMouseEventNearestFeature } from '../../../../utils/mapboxHelper';
 import {
   POINTS_LAYER_ID,
@@ -26,6 +25,8 @@ import {
 import { TrackEditionState } from './types';
 import { getNewLine } from './utils';
 import { entityDoUpdate } from '../../components/LinearMetadata';
+import { Tool } from '../editorContextTypes';
+import { DEFAULT_COMMON_TOOL_STATE } from '../commonToolState';
 
 const TrackEditionTool: Tool<TrackEditionState> = {
   id: 'track-edition',
